moved outside click into LayerButton
